Finalized full functioning first page
Still date limiter is pending and showing data for one location in Sydney, CBD.
Will use fused location provider for that.
In this version, added blow
Added Icon of a Venue and cashing, resizing with Glide
Ensure data persist for a query in Room database smoothly
Currently showing Contacts, category, location and Icons of a venue

+++ app/src/main/java/mohsin/reza/whatsaround/vo/Location.java
@@ -24,9 +24,10 @@ public class Location {
     @SerializedName("lng")
     public final double lng;
     @SerializedName("distance")
-    public final double distance;
+    public final Integer distance;
 
-    public Location(String address, String crossStreet, String city, String state, String postalCode, String country, Double lat, Double lng, double distance)
+    public Location(String address, String crossStreet, String city, String state, String postalCode, String country,
+                    Double lat, Double lng, Integer distance)
     {
         this.address=address;
         this.crossStreet=crossStreet;
